requirements.txt
RPi.GPIO==0.6.1
decorator==4.0.6
ipython==4.0.3
ipython-genutils==0.1.0
path.py==8.1.2
pexpect==4.0.1
pickleshare==0.6
ptyprocess==0.5
python-espeak==0.5
simplegeneric==0.8.1
traitlets==4.1.0
